Fixed delete user - need to delete only calendars the user created, need to delete the user from other's calendars

+++ ServiceModel/CalendarService.cs
@@ -40,7 +40,10 @@ namespace ServiceModel
             {
                 foreach (Calendar calendar in user.Calendars)
                 {
-                    DeleteCalendar(calendar);
+                    if (calendar.Creator.ID == user.ID) // only if he is the creator
+                    {
+                        DeleteCalendar(calendar);
+                    }
                 }
             }
             UserDB userDB = new UserDB();
